refactor: split gnosis pay auth component
