+++ frontend/app/src/modules/external-services/gnosis-pay/components/GnosisPayAuthStep.vue
