+++ frontend/app/src/modules/history/refresh/HistoryRefreshProtocolEvents.vue
@@ -3,7 +3,7 @@ import { getTextToken, toHumanReadable } from '@rotki/common';
 import { get, set } from '@vueuse/core';
 import { isEqual } from 'es-toolkit';
 import { useExternalApiKeys } from '@/composables/settings/api-keys/external';
-import { useMoneriumOAuth } from '@/composables/settings/api-keys/external/monerium-oauth';
+import { useMoneriumOAuth } from '@/modules/external-services/monerium/use-monerium-auth';
 import { OnlineHistoryEventsQueryType } from '@/types/history/events/schemas';
 
 const modelValue = defineModel<OnlineHistoryEventsQueryType[]>({ required: true });
